requirements.txt
aiomysql==0.3.2
aiosqlite==0.21.0
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
black==25.12.0
bleach==6.3.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
coverage==7.13.0
cryptography==46.0.3
ecdsa==0.19.1
fastapi==0.124.0
filelock==3.20.0
flake8==7.3.0
fsspec==2025.12.0
greenlet==3.3.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
hypothesis==6.148.7
idna==3.11
iniconfig==2.3.0
isort==7.0.0
Jinja2==3.1.6
joblib==1.5.2
kaggle==1.8.2
kagglesdk==0.1.13
librt==0.7.3
Mako==1.3.10
MarkupSafe==3.0.3
mccabe==0.7.0
mpmath==1.3.0
mypy==1.19.0
mypy_extensions==1.1.0
networkx==3.6.1
numpy==2.3.5
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.1
pluggy==1.6.0
protobuf==6.33.2
pyasn1==0.6.1
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
PyMySQL==1.1.2
pytest==9.0.2
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-slugify==8.0.4
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
setuptools==80.9.0
six==1.17.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.44
starlette==0.50.0
structlog==25.5.0
sympy==1.14.0
text-unidecode==1.3
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
types-requests==2.32.4.20250913
types-tqdm==4.67.0.20250809
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.1
uvicorn==0.38.0
webencodings==0.5.1
